Assume the user knows best, don't sanitize remapped names
Previously the sanitizer was run on user remapped names, as the
sanitizer rewrites perfectly legal git names (such as __.*) this is
probably not what the user wants. Instead assume the user knows best
and just send the remapped value on to fast-import.

This change will break existing conversions which rely on the
post-remap sanitizer.